refactor: Split skill and allowlist logic into runtime and exec modules
