+++ cmd/app/newTUI.go
@@ -20,7 +20,6 @@ import (
 	telegramTool "github.com/pardnchiu/agenvoy/internal/runtime/telegram/tool"
 	"github.com/pardnchiu/agenvoy/internal/runtime/tui"
 	"github.com/pardnchiu/agenvoy/internal/session"
-	"github.com/pardnchiu/agenvoy/internal/skill"
 	"github.com/pardnchiu/agenvoy/internal/tools/agent/subagent"
 	go_pkg_sandbox "github.com/pardnchiu/go-pkg/sandbox"
 )
@@ -69,7 +68,7 @@ func newTUI() {
 	defer mcpManager.Close()
 
 	registry := buildAgentRegistry()
-	scanner := skill.NewScanner()
+	scanner := runtime.NewSkillScanner()
 	selectorBot := plannerSelector(registry)
 
 	host.Set(selectorBot, registry, scanner)
