+++ internal/agents/exec/allow/skill/match.go
@@ -1,0 +1,17 @@
+package allowSkill
+
+import "strings"
+
+func Match(workDir, name string) bool {
+	name = strings.TrimSpace(name)
+	if name == "" {
+		return false
+	}
+	if LoadGlobal()[name] {
+		return true
+	}
+	if strings.TrimSpace(workDir) != "" && LoadProject(workDir)[name] {
+		return true
+	}
+	return false
+}
